backend/requirements.txt
# Core Django
Django
djangorestframework
django-filter
django-cors-headers
logger
# Authentication
djangorestframework-simplejwt

# API Documentation
drf-spectacular
drf_yasg



#similarity check with viral post
sentence-transformers 
faiss-cpu
torch

# Database - MongoDB
mongoengine
pymongo
dnspython
motor

# Task Queue
celery
django-celery-beat
django-celery-results
redis
django-redis

# Storage
django-storages
boto3
awscli
# AI/ML
openai
transformers
torch
sentence-transformers
qdrant-client
odfpy


# # Development
# django-debug-toolbar
# black
# flake8
# isort
# mypy

# Production
gunicorn
uvicorn[standard]
supervisor
sentry-sdk

# Utilities
python-dotenv
Pillow
requests
python-dateutil
pytz


#agentic library
 nest_asyncio
 uvicorn
 semantic_kernel
 ollama
 asyncio
 google-generativeai

#excel reading
pandas
xlrd==2.0.1
xlwt
openpyxl

#webscraping
selenium 
webdriver-manager


# Ubuntu/Debian
# sudo apt-get update
# sudo apt-get install -y chromium-browser chromium-chromedriver

# # Or use Google Chrome
# wget https://dl.google.com/linux/direct/google-chrome-stable_current_amd64.deb
# sudo dpkg -i google-chrome-stable_current_amd64.deb
# sudo apt-get install -f














# # Django
# Django==4.2.7
# django-cors-headers==4.3.1
# django-filter==23.5
# python-dotenv==1.0.0

# # Django REST Framework
# djangorestframework==3.14.0
# djangorestframework-simplejwt==5.3.1

# # MongoDB Support
# djongo==1.3.6
# pymongo==4.6.0
# sqlparse==0.4.4
# dnspython==2.4.2
# certifi==2023.11.17

# # API Documentation
# drf-spectacular==0.27.0

# # Celery & Redis
# celery==5.3.4
# redis==5.0.1
# django-redis==5.4.0
# django-celery-beat==2.5.0
# django-celery-results==2.5.1

# # File Storage (MinIO/S3)
# boto3==1.34.10
# django-storages==1.14.2

# # Development Tools
# django-debug-toolbar==4.2.0

# # HTTP & Requests
# requests==2.31.0
# urllib3==2.1.0

# # Utilities
# Pillow==10.1.0
# python-dateutil==2.8.2
# pytz==2023.3